LayoutTests/imported/w3c/web-platform-tests/css/css-anchor-position/position-anchor-003-expected.txt


FAIL Layout is updated on `position-anchor` changes assert_equals: expected 200 but got 0
